Correction

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	int i = 0;
 	unsigned int pf = 2u;
-	long long int n = 612852475143;
+	long int n = 612852475143;
 
 	while (1)
 	{
